pull data

+++ src/app/utls/actions/delete/deleteBlogs.ts
@@ -10,6 +10,10 @@ const deleteBlogs = async (id: string) => {
     try {
         const session = await getServerSession(authOptions);
 
+        if (session?.user?.role !== "Admin") {
+            throw new Error("Forbidden: Only admins can delete");
+        }
+
         const res = await fetch(`${process.env.BASE_URL}/blogs/${id}`, {
             method: "DELETE",
             headers: {
